[STREAMPIPES-642] 1-class processor; misc changes structure and ressource

+++ streampipes-model/src/main/java/org/apache/streampipes/model/runtime/Event.java
@@ -189,7 +189,7 @@ public class Event {
     }
   }
 
-  public void addField(String runtimeName, Float value) {
+  public void addField(String runtimeName, double value) {
     addPrimitive(runtimeName, value);
   }
